Rename random_generation.py.
